Move RoofDualityComposite into its own submodule
