+++ dimod/reference/composites/__init__.py
@@ -14,8 +14,9 @@
 #
 # ================================================================================================
 
-from dimod.reference.composites.fixedvariable import *
+from dimod.reference.composites.fixedvariable import FixedVariableComposite
 from dimod.reference.composites.higherordercomposites import *
+from dimod.reference.composites.roofduality import RoofDualityComposite
 from dimod.reference.composites.scalecomposite import *
 from dimod.reference.composites.spin_transform import *
 from dimod.reference.composites.structure import *
