+++ dimod/reference/composites/fixedvariable.py
@@ -12,22 +12,18 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 #
-# ================================================================================================
+# =============================================================================
 """
 A composite that fixes the variables provided and removes them from the
 bqm object before sending to its child sampler.
-
-See `Ocean Glossary <https://docs.ocean.dwavesys.com/en/latest/glossary.html>`_ for explanations
-of technical terms in descriptions of Ocean tools.
 """
 import numpy as np
 
 from dimod.core.composite import ComposedSampler
-from dimod.roof_duality import fix_variables
 from dimod.sampleset import SampleSet
 
 
-__all__ = 'FixedVariableComposite', 'RoofDualityComposite'
+__all__ = ['FixedVariableComposite']
 
 
 class FixedVariableComposite(ComposedSampler):
@@ -174,55 +170,3 @@ def _release_response(response, fixed_variables):
 
     return SampleSet(data, original_variables, response.info,
                      response.vartype)
-
-
-class RoofDualityComposite(FixedVariableComposite):
-    """Uses roof duality to assign some variables before invoking child sampler.
-
-    Uses the :func:`~dimod.roof_duality.fix_variables` function to determine
-    variable assignments, then fixes them before calling the child sampler.
-    Returned samples include the fixed variables.
-
-    Args:
-       child (:obj:`dimod.Sampler`):
-            A dimod sampler. Used to sample the bqm after variables have been
-            fixed.
-
-    See also:
-        :func:`~dimod.roof_duality.fix_variables` for a description of the
-        algorithm.
-
-    """
-
-    @property
-    def parameters(self):
-        params = self.child.parameters.copy()
-        params['sampling_mode'] = []
-        return params
-
-    def sample(self, bqm, sampling_mode=True, **parameters):
-        """Sample from the provided binary quadratic model.
-
-        Uses the :func:`~dimod.roof_duality.fix_variables` function to determine
-        which variables to fix.
-
-        Args:
-            bqm (:obj:`dimod.BinaryQuadraticModel`):
-                Binary quadratic model to be sampled from.
-
-            sampling_mode (bool, optional, default=True):
-                In sampling mode, only roof-duality is used. When
-                `sampling_mode` is false, strongly connected components are used
-                to fix more variables, but in some optimal solutions these
-                variables may take different values.
-
-            **parameters:
-                Parameters for the child sampler.
-
-        Returns:
-            :obj:`dimod.SampleSet`
-
-        """
-        # use roof-duality to decide which variables to fix
-        parameters['fixed_variables'] = fix_variables(bqm, sampling_mode=sampling_mode)
-        return super(RoofDualityComposite, self).sample(bqm, **parameters)
